pyproject.toml
[project]
name = "doctor-appointment-api"
version = "0.1.0"
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy[asyncio]",
  "asyncpg",
  "passlib[bcrypt]",
  "python-jose[cryptography]",
  "pydantic[email]",
  "pytest",
  "httpx"
]
